pyproject.toml
[tool.poetry]
name = "aider-vertex"
version = "1.1.5"
description = "Nix wrapper for Aider (v0.86.1) with Vertex AI support"
authors = ["Andrey Stepantsov <andrey.stepantosv@gmail.com>"]
readme = "README.md"

# This ensures the build backend finds your code in the subdirectory
packages = [
    { include = "aider_vertex" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
aider-chat = "^0.86.1"
google-cloud-aiplatform = "1.115.0"
google-genai = "1.2.0"
# These are often required as sub-dependencies for Vertex AI
google-api-core = "*"
google-cloud-core = "*"
google-cloud-storage = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# This creates the 'aider-vertex' command when you install the package
[tool.poetry.scripts]
aider-vertex = "aider_vertex.main:main"
